requirements.txt
boto3==1.34.0
botocore==1.34.162
dnspython==2.6.0
numpy==1.26.4
pandas==2.2.0
pymongo==4.6.1
python-dotenv==1.2.1
openpyxl==3.1.2
urllib3==2.6.3